<fix> Solucion del merge votacion
Se han solventado los errores creados a la hora de mergear los cambios de todas las ramas.
Para una mejor visualización y gestión del código se han reorganizado los test referentes a Selenium en el archivo test-Selenium,
 eliminando el anterior archivo test-customurls. Este commit se ha realizado en equipo, trabajando paralelamente.

<tarea:Errores en el Merge de los equipos #234>
