+++ decide/voting/tests.py
@@ -36,7 +36,7 @@ import time
 import datetime
 
 
-class VotingTestCase(APITestCase):
+class VotingTestCase(BaseTestCase):
 
     def setUp(self):
         super().setUp()
@@ -491,338 +491,6 @@ class VotingTestCase(APITestCase):
         self.assertEqual(response.status_code, 400)
         self.assertEqual(response.json(), 'Voting has not being tallied')        
         
-        
-class EndDateTestCase(StaticLiveServerTestCase):
-
-    def setUp(self):
-        self.base=BaseTestCase()
-        self.base.setUp()
-
-        options = webdriver.ChromeOptions()
-        options.headless = True
-        self.driver = webdriver.Chrome(options=options)
-        self.vars = {}
-
-        super().setUp()
-
-    def wait_for_window(self, timeout = 2):
-        time.sleep(round(timeout / 1000))
-        wh_now = self.driver.window_handles
-        wh_then = self.vars["window_handles"]
-        if len(wh_now) > len(wh_then):
-            return set(wh_now).difference(set(wh_then)).pop()      
-
-    #Test de selenium de la fecha de finalizacion caso positivo
-    def test_end_date_p(self):
-        driver = self.driver
-        
-        #Iniciar sesion
-        driver.get("http://localhost:8000/admin/login/?next=/admin/")
-        time.sleep(1)
-        self.driver.find_element_by_id('id_username').send_keys("practica")
-        time.sleep(1)
-        self.driver.find_element_by_id('id_password').send_keys("practica", Keys.ENTER)
-        time.sleep(1)
-
-        #Crear auths
-        self.driver.find_element_by_link_text("Auths").click()
-        self.driver.find_element_by_link_text("AÑADIR AUTH").click()
-        self.driver.find_element(By.ID, "id_name").click()
-        self.driver.find_element(By.ID, "id_name").send_keys("localhost")
-        self.driver.find_element(By.ID, "id_url").click()
-        self.driver.find_element(By.ID, "id_url").send_keys("http://localhost:8000")
-        self.driver.find_element(By.ID, "id_me").click()
-        self.driver.find_element(By.NAME, "_save").click()
-
-
-        #Crear votacion
-        self.driver.get("http://localhost:8000/admin/")
-        driver.find_element_by_link_text("Votings").click()
-        driver.find_element_by_link_text("AÑADIR VOTING").click()
-        self.driver.find_element(By.ID, "id_name").click()
-        self.driver.find_element(By.ID, "id_name").send_keys("prueba")
-        self.driver.find_element(By.ID, "id_desc").click()
-        self.driver.find_element(By.ID, "id_desc").send_keys("prueba")
-
-        #Crear question dentro de crear votacion
-        self.driver.find_element(By.CSS_SELECTOR, ".field-question .related-widget-wrapper").click()
-        self.vars["window_handles"] = self.driver.window_handles
-        self.driver.find_element(By.CSS_SELECTOR, "#add_id_question > img").click()
-        self.vars["win8304"] = self.wait_for_window(2000)
-        self.vars["root"] = self.driver.current_window_handle
-        self.driver.switch_to.window(self.vars["win8304"])
-        self.driver.find_element(By.ID, "id_desc").send_keys("prueba")
-        self.driver.find_element(By.ID, "id_options-0-number").send_keys("1")
-        self.driver.find_element(By.ID, "id_options-0-number").click()
-        self.driver.find_element(By.ID, "id_options-1-number").send_keys("2")
-        self.driver.find_element(By.ID, "id_options-1-number").click()
-        self.driver.find_element(By.ID, "id_options-0-option").click()
-        self.driver.find_element(By.ID, "id_options-0-option").send_keys("prueba 1")
-        self.driver.find_element(By.ID, "id_options-1-option").click()
-        self.driver.find_element(By.ID, "id_options-1-option").send_keys("prueba 2")
-        self.driver.find_element(By.NAME, "_save").click()
-        self.driver.switch_to.window(self.vars["root"])
-        self.driver.find_element(By.LINK_TEXT, "Hoy").click()
-        self.driver.find_element(By.ID, "id_end_date_1").click()
-        self.driver.find_element(By.ID, "id_end_date_1").send_keys("23:50:00")
-
-        #Seleccionar auth
-        dropdown = self.driver.find_element(By.ID, "id_auths")
-        dropdown.find_element(By.XPATH, "//option[. = 'http://localhost:8000']").click()
-
-        #Guardar la votacion
-        self.driver.find_element(By.NAME, "_save").click()
-        time.sleep(1)
-        self.driver.find_element(By.CSS_SELECTOR, ".row1:nth-child(1) > .field-name").get_attribute('value')
-        self.assertEqual("prueba", driver.find_element(By.CSS_SELECTOR, ".row1:nth-child(1) > .field-name").text)
-        time.sleep(1)
-
-        #Iniciar la votacion
-        self.driver.find_element(By.NAME, "_selected_action").click()
-        dropdown = self.driver.find_element(By.NAME, "action")
-        dropdown.find_element(By.XPATH, "//option[. = 'Start']").click()
-        element = self.driver.find_element(By.NAME, "action")
-        actions = ActionChains(self.driver)
-        actions.move_to_element(element).click_and_hold().perform()
-        element = self.driver.find_element(By.NAME, "action")
-        actions = ActionChains(self.driver)
-        actions.move_to_element(element).perform()
-        element = self.driver.find_element(By.NAME, "action")
-        actions = ActionChains(self.driver)
-        actions.move_to_element(element).release().perform()
-        self.driver.find_element(By.NAME, "action").click()
-        self.driver.find_element(By.NAME, "index").click()
-
-        #Crear censo
-        driver.get("http://localhost:8000/admin/login/?next=/admin/")
-        driver.find_element_by_link_text("Censuss").click()
-        driver.find_element_by_link_text("AÑADIR CENSUS").click()
-        self.driver.find_element(By.ID, "id_voting_id").click()
-        self.driver.find_element(By.ID, "id_voting_id").send_keys("1")
-        self.driver.find_element(By.ID, "id_voter_id").click()
-        self.driver.find_element(By.ID, "id_voter_id").send_keys("1")
-        self.driver.find_element(By.NAME, "_save").click()
-
-        #Realizar la votacion
-        driver.get("http://localhost:8000/booth/1")
-        time.sleep(1)
-        self.driver.find_element(By.ID, "username").click()
-        time.sleep(1)
-        self.driver.find_element(By.ID, "username").send_keys("practica")
-        time.sleep(1)
-        self.driver.find_element(By.ID, "password").click()
-        time.sleep(1)
-        self.driver.find_element(By.ID, "password").send_keys("practica")
-        time.sleep(1)
-        self.driver.find_element(By.ID, "password").send_keys(Keys.ENTER)
-        time.sleep(1)
-        self.driver.find_element(By.CSS_SELECTOR, "#\\__BVID__12 .custom-control-label").click()
-        time.sleep(1)
-        self.driver.find_element(By.CSS_SELECTOR, ".btn:nth-child(4)").click()
-        time.sleep(1)
-        self.driver.find_element(By.LINK_TEXT, "Logout").click()
-
-    #Test de selenium de la fecha de finalizacion caso negativo
-    def test_end_date_n(self):
-        driver = self.driver
-
-        #Iniciar sesion
-        driver.get("http://localhost:8000/admin/login/?next=/admin/")
-        time.sleep(1)
-        self.driver.find_element_by_id('id_username').send_keys("practica")
-        time.sleep(1)
-        self.driver.find_element_by_id('id_password').send_keys("practica", Keys.ENTER)
-        time.sleep(1)
-
-        #Crear auths
-        self.driver.find_element_by_link_text("Auths").click()
-        self.driver.find_element_by_link_text("AÑADIR AUTH").click()
-        self.driver.find_element(By.ID, "id_name").click()
-        self.driver.find_element(By.ID, "id_name").send_keys("localhost")
-        self.driver.find_element(By.ID, "id_url").click()
-        self.driver.find_element(By.ID, "id_url").send_keys("http://localhost:8000")
-        self.driver.find_element(By.ID, "id_me").click()
-        self.driver.find_element(By.NAME, "_save").click()
-
-        #Crear votacion
-        driver.get("http://localhost:8000/admin/login/?next=/admin/")
-        driver.find_element_by_link_text("Votings").click()
-        driver.find_element_by_link_text("AÑADIR VOTING").click()
-        self.driver.find_element(By.ID, "id_name").click()
-        self.driver.find_element(By.ID, "id_name").send_keys("prueba")
-        self.driver.find_element(By.ID, "id_desc").click()
-        self.driver.find_element(By.ID, "id_desc").send_keys("prueba")
-
-        #Crear question
-        self.driver.find_element(By.CSS_SELECTOR, ".field-question .related-widget-wrapper").click()
-        self.vars["window_handles"] = self.driver.window_handles
-        self.driver.find_element(By.CSS_SELECTOR, "#add_id_question > img").click()
-        self.vars["win8304"] = self.wait_for_window(2000)
-        self.vars["root"] = self.driver.current_window_handle
-        self.driver.switch_to.window(self.vars["win8304"])
-        self.driver.find_element(By.ID, "id_desc").send_keys("prueba")
-        self.driver.find_element(By.ID, "id_options-0-number").send_keys("1")
-        self.driver.find_element(By.ID, "id_options-0-number").click()
-        self.driver.find_element(By.ID, "id_options-1-number").send_keys("2")
-        self.driver.find_element(By.ID, "id_options-1-number").click()
-        self.driver.find_element(By.ID, "id_options-0-option").click()
-        self.driver.find_element(By.ID, "id_options-0-option").send_keys("prueba 1")
-        self.driver.find_element(By.ID, "id_options-1-option").click()
-        self.driver.find_element(By.ID, "id_options-1-option").send_keys("prueba 2")
-        self.driver.find_element(By.NAME, "_save").click()
-        self.driver.switch_to.window(self.vars["root"])
-        self.driver.find_element(By.LINK_TEXT, "Hoy").click()
-        self.driver.find_element(By.ID, "id_end_date_1").click()
-        self.driver.find_element(By.ID, "id_end_date_1").send_keys("00:10:00")
-
-        #Seleccionar auth
-        dropdown = self.driver.find_element(By.ID, "id_auths")
-        dropdown.find_element(By.XPATH, "//option[. = 'http://localhost:8000']").click()
-
-        #Guardar la votacion
-        self.driver.find_element(By.NAME, "_save").click()
-        
-        #Comprobar que no se ha aceptado el guardado y sigue en la pagina de creacion
-        p = driver.find_element(By.ID, "id_name")
-        self.assertEqual("prueba", p.get_attribute('value'))
-        d = driver.find_element(By.ID, "id_desc")
-        self.assertEqual("prueba", d.get_attribute('value'))
-
-        #Cerrar sesion
-        self.driver.find_element(By.CSS_SELECTOR, "a:nth-child(4)").click()    
-
-    def tearDown(self):           
-        super().tearDown()
-        self.driver.quit()
-        self.base.tearDown()
-
-class VotingLocalSaveTestCase(StaticLiveServerTestCase):
-
-    def setUp(self):
-        self.base=BaseTestCase()
-        self.base.setUp()
-
-        options = webdriver.ChromeOptions()
-        options.headless = True
-        self.driver = webdriver.Chrome(options=options)
-        super().setUp()
-
-
-
-    # Test selenium de guardar
-    def test_guardar(self):
-        driver=self.driver
-        User.objects.create_superuser('egcVotacion','votacion@decide.com','egcVotacion')
-
-        # Login
-        self.driver.get("http://localhost:8000/admin/login/?next=/admin/")
-        self.driver.find_element_by_id("id_username").send_keys("egcVotacion")
-        self.driver.find_element_by_id("id_password").send_keys("egcVotacion")
-        self.driver.find_element_by_id("id_password").send_keys(Keys.ENTER)
-
-
-        # Añado pregunta
-        self.driver.find_element_by_link_text("Questions").click()
-        self.driver.find_element_by_link_text("ADD QUESTION").click()
-        self.driver.find_element(By.ID, "id_desc").click()
-        self.driver.find_element(By.ID, "id_desc").send_keys("pregunta de prueba")
-        self.driver.find_element_by_id("id_options-0-option").click()
-        self.driver.find_element_by_id("id_options-0-option").send_keys("opcion1")
-        self.driver.find_element_by_id("id_options-0-option").click()
-        self.driver.find_element(By.ID, "id_options-1-option").send_keys("opcion2")
-        self.driver.find_element(By.NAME, "_save").click()
-
-        # Añado auth
-        self.driver.get("http://localhost:8000/admin/")
-        self.driver.find_element_by_link_text("Auths").click()
-        self.driver.find_element_by_link_text("ADD AUTH").click()
-        self.driver.find_element(By.ID, "id_name").click()
-        self.driver.find_element(By.ID, "id_name").send_keys("localhost")
-        self.driver.find_element(By.ID, "id_url").click()
-        self.driver.find_element(By.ID, "id_url").send_keys("http://localhost:8000")
-        self.driver.find_element(By.ID, "id_me").click()
-        self.driver.find_element(By.NAME, "_save").click()
-
-        # Añado votacion
-        self.driver.get("http://localhost:8000/admin/")
-        self.driver.find_element_by_link_text("Votings").click()
-        self.driver.find_element_by_link_text("ADD VOTING").click()
-        self.driver.find_element(By.ID, "id_name").send_keys("prueba")
-        self.driver.find_element(By.ID, "id_desc").send_keys("descripcion de prueba")
-        dropdown = self.driver.find_element(By.NAME, "question")
-        dropdown.find_element(By.XPATH, "//option[. = 'pregunta de prueba']").click()
-        Select(driver.find_element(By.ID,"id_auths")).select_by_visible_text("http://localhost:8000")
-        self.driver.find_element(By.NAME, "_save").click()
-
-
-        # Inicio votacion
-        self.driver.get("http://localhost:8000/admin/")
-        self.driver.find_element_by_link_text("Votings").click()
-        self.driver.find_element(By.NAME, "_selected_action").click()
-        dropdown = self.driver.find_element(By.NAME, "action")
-        dropdown.find_element(By.XPATH, "//option[. = 'Start']").click()
-        self.driver.find_element(By.NAME, "index").click()
-
-        #Termino votacion
-        self.driver.find_element(By.NAME, "_selected_action").click()
-        dropdown = self.driver.find_element(By.NAME, "action")
-        dropdown.find_element(By.XPATH, "//option[. = 'Stop']").click()
-        self.driver.find_element(By.NAME, "index").click()
-
-        #Recuento de votacion
-        self.driver.find_element(By.NAME, "_selected_action").click()
-        dropdown = self.driver.find_element(By.NAME, "action")
-        dropdown.find_element(By.XPATH, "//option[. = 'Tally']").click()
-        self.driver.find_element(By.NAME, "index").click()
-       
-        # Accedo a la votacion 
-        self.driver.find_element(By.LINK_TEXT, "prueba").click()
-        # Compruebo que está vacío el campo
-        self.assertEqual("",self.driver.find_element(By.CSS_SELECTOR, ".field-file .readonly").text)
-        
-        # Guardo la votacion
-        self.driver.find_element(By.LINK_TEXT, "Votings").click()
-        self.driver.find_element(By.NAME, "_selected_action").click()
-        dropdown = self.driver.find_element(By.NAME, "action")
-        dropdown.find_element(By.XPATH, "//option[. = 'Save']").click()
-        self.driver.find_element(By.NAME, "action").click()
-        self.driver.find_element(By.NAME, "index").click()
-        
-        # Accedo a la votacion
-        self.driver.find_element(By.LINK_TEXT, "prueba").click()
-        # Compruebo que se ha guardado
-        self.assertNotEqual("",self.driver.find_element(By.CSS_SELECTOR, ".field-file .readonly").text)
-
-        # Borro la votacion
-        self.driver.find_element(By.LINK_TEXT,"Delete").click()
-        self.driver.find_element(By.XPATH,"//input[@value=\"Yes, I'm sure\"]").click()
-
-        # Borro la pregunta
-        self.driver.get("http://localhost:8000/admin/voting/")
-        self.driver.find_element_by_link_text("Questions").click()
-        self.driver.find_element(By.NAME, "_selected_action").click()
-        dropdown = self.driver.find_element(By.NAME, "action").click()
-        self.driver.find_element(By.XPATH, "//option[. = 'Delete selected questions']").click()
-        self.driver.find_element(By.NAME, "action").click()
-        self.driver.find_element(By.NAME, "index").click()
-        self.driver.find_element(By.XPATH,"//input[@value=\"Yes, I'm sure\"]").click()
-
-        # Borro el auth
-        self.driver.get("http://localhost:8000/admin/base/auth/")
-        self.driver.find_element(By.NAME, "_selected_action").click()
-        dropdown = self.driver.find_element(By.NAME, "action").click()
-        self.driver.find_element(By.XPATH, "//option[. = 'Delete selected auths']").click()
-        self.driver.find_element(By.NAME, "action").click()
-        self.driver.find_element(By.NAME, "index").click()
-        self.driver.find_element(By.XPATH,"//input[@value=\"Yes, I'm sure\"]").click()
-
-    def tearDown(self):
-        super().tearDown()
-        self.driver.quit()
-
-        self.base.tearDown()
-
-# Yes/No question model test
 class YesNoQuestionModelTestCase(BaseTestCase):
     def setUp(self):
         super().setUp()
@@ -879,139 +547,6 @@ class YesNoQuestionModelTestCase(BaseTestCase):
         self.assertEquals(v.question.yes_no_question, True)
         self.assertEquals(v.name, 'Yes/No voting test')
         
-# Yes/No question view test
-class YesNoQuestionViewTestCase(StaticLiveServerTestCase):
-
-    def setUp(self):
-        #Load base test functionality for decide
-        self.base = BaseTestCase()
-        self.base.setUp()
-
-        options = webdriver.ChromeOptions()
-        options.headless = True
-        self.driver = webdriver.Chrome(options=options)
-
-        super().setUp()     
-
-    def tearDown(self):
-        super().tearDown()
-        self.driver.quit()
-
-        self.base.tearDown()
-
-    # Verify if a Yes/No question is created correctly
-    def test_create_yes_no_question(self):       
-        driver = self.driver
-        # Creation of a super user to being able to create questions
-        User.objects.create_superuser('egcVotacion', 'votacion@decide.com', 'egcVotacion')
-        self.driver.get(f'{self.live_server_url}/admin/')
-        # Web log in
-        self.driver.find_element_by_id('id_username').send_keys("egcVotacion")
-        self.driver.find_element_by_id('id_password').send_keys("egcVotacion", Keys.ENTER)
-        time.sleep(3)
-
-
-        # Access to add question form
-        driver.find_element_by_link_text("Questions").click()
-        time.sleep(1)
-        driver.find_element_by_link_text("Add question").click()
-        time.sleep(1)
-
-        # Creation of a yes/no question example
-        driver.find_element_by_id('id_desc').send_keys("Si/No prueba")
-        driver.find_element_by_xpath("//form[@id='question_form']/div/fieldset/div[3]/div/label").click()
-        driver.find_element_by_name("_save").click()
-        time.sleep(3)
-
-        # Verification of the creation of the question
-        driver.find_element_by_xpath("(//a[contains(text(),'Si/No prueba')])[2]").click()
-        time.sleep(2)
-        self.assertEqual("Si/No prueba", driver.find_element_by_xpath("//textarea[@id='id_desc']").text)
-        self.assertEqual("YES", driver.find_element_by_id("id_options-0-option").text)
-        self.assertEqual("NO", driver.find_element_by_id("id_options-1-option").text)
-
-
-    # Verify if the extra options are not save in a Yes/No question
-    def test_create_yes_no_question_extra(self):
-        driver = self.driver
-         # Creation of a super user to being able to create questions
-        User.objects.create_superuser('egcVotacion', 'votacion@decide.com', 'egcVotacion')
-        self.driver.get(f'{self.live_server_url}/admin/')
-        # Web log in
-        self.driver.find_element_by_id('id_username').send_keys("egcVotacion")
-        self.driver.find_element_by_id('id_password').send_keys("egcVotacion", Keys.ENTER)
-        time.sleep(3)
-
-        # Access to add question form
-        driver.find_element_by_link_text("Questions").click()
-        time.sleep(1)
-        driver.find_element_by_link_text("Add question").click()
-        time.sleep(1)
-
-        # Creation of a yes/no question example with extra answers
-        driver.find_element_by_id('id_desc').send_keys("Si/No prueba con extra")
-        driver.find_element_by_xpath("//form[@id='question_form']/div/fieldset/div[3]/div/label").click()
-        driver.find_element_by_id("id_options-0-option").send_keys("Primer extra")
-        driver.find_element_by_id("id_options-1-option").send_keys("Segundo extra")
-        driver.find_element_by_name("_save").click()
-        time.sleep(3)
-
-        # Verification of the creation of the question without the extra answers
-        driver.find_element_by_xpath("(//a[contains(text(),'Si/No prueba con extra')])[2]").click()
-        self.assertEqual("Si/No prueba con extra", driver.find_element_by_xpath("//textarea[@id='id_desc']").text)
-        self.assertEqual("YES", driver.find_element_by_id("id_options-0-option").text)
-        self.assertEqual("NO", driver.find_element_by_id("id_options-1-option").text)
-        self.assertEqual("", driver.find_element_by_id("id_options-2-option").text)
-        self.assertEqual("", driver.find_element_by_id("id_options-3-option").text)
-
-    def test_create_yes_no_voting(self):
-        driver = self.driver
-        # Creation of a super user to being able to create questions
-        User.objects.create_superuser('egcVotacion', 'votacion@decide.com', 'egcVotacion')
-        driver.get(f'{self.live_server_url}/admin/')
-        # Web log in
-        driver.find_element_by_id('id_username').send_keys("egcVotacion")
-        driver.find_element_by_id('id_password').send_keys("egcVotacion", Keys.ENTER)
-        time.sleep(3)
-
-        # Access to add question form
-        driver.find_element_by_link_text("Questions").click()
-        time.sleep(1)
-        driver.find_element_by_link_text("Add question").click()
-        time.sleep(1)
-
-        # Creation of a yes/no question example
-        driver.find_element_by_id('id_desc').send_keys("Si/No prueba")
-        driver.find_element_by_xpath("//form[@id='question_form']/div/fieldset/div[3]/div/label").click()
-        driver.find_element_by_name("_save").click()
-        time.sleep(3)
-
-       # Creation of auth
-        driver.get(f'{self.live_server_url}/admin/')
-        driver.find_element_by_link_text("Auths").click()
-        driver.find_element_by_link_text("Add auth").click()
-        driver.find_element(By.ID, "id_name").click()
-        driver.find_element(By.ID, "id_name").send_keys("localhost")
-        driver.find_element(By.ID, "id_url").click()
-        driver.find_element(By.ID, "id_url").send_keys("http://localhost:8000")
-        driver.find_element(By.ID, "id_me").click()
-        driver.find_element_by_name("_save").click()
-
-        # Creation of a yes/no voting example
-        driver.get(f'{self.live_server_url}/admin/')
-        driver.find_element_by_link_text("Votings").click()
-        driver.find_element_by_link_text("Add voting").click()
-        driver.find_element(By.ID, "id_name").send_keys("Votacion de Si/No")
-        driver.find_element(By.ID, "id_desc").send_keys("votacion con preguntas de Si")
-        self.dropdown = driver.find_element(By.NAME, "question")
-        self.dropdown.find_element(By.XPATH, "//option[. = 'Si/No prueba']").click()
-        self.dropdown = driver.find_element(By.NAME, "auths")
-        self.dropdown.find_element(By.XPATH, "//option[. = 'http://localhost:8000']").click()
-        driver.find_element_by_name("_save").click()
-        self.assertEqual("Votacion de Si/No", driver.find_element_by_xpath("(//a[contains(text(),'Votacion de Si/No')])[2]").text)
-
-
-#TESTS MODELO FLOR
 class VotingModelTestCase(BaseTestCase):
     def setUp(self):
         q=Question(desc="Description")
@@ -1133,150 +668,3 @@ class VotacionPrefModelTestCase2(BaseTestCase):
         self.assertNotEquals(vp.question.options.all()[1].number,3)
         self.assertNotEquals(vp.question.options.all()[2].option,"ave")
         self.assertNotEquals(vp.question.options.all()[2].number,1)
-
-
-#TEST CON SELENIUM MODELO FLOR
-class TestQues(StaticLiveServerTestCase):
-    def setUp(self):
-        self.driver = webdriver.Firefox()
-        self.driver.implicitly_wait(30)
-        self.base_url = "https://www.google.com/"
-        self.verificationErrors = []
-        self.accept_next_alert = True
-    
-    def test_ques(self):
-        driver = self.driver
-        driver.get("http://localhost:8000/admin/login/?next=/admin/")
-        driver.find_element_by_id("id_username").click()
-        driver.find_element_by_id("id_username").clear()
-        driver.find_element_by_id("id_username").send_keys("admin")
-        driver.find_element_by_id("id_password").click()
-        driver.find_element_by_id("id_password").clear()
-        driver.find_element_by_id("id_password").send_keys("flocorlop")
-        driver.find_element_by_xpath("//input[@value='Log in']").click()
-        driver.find_element_by_link_text("Questions").click()
-        driver.find_element_by_class_name("addlink").click()
-        driver.find_element_by_id("id_desc").click()
-        driver.find_element_by_id("id_desc").clear()
-        driver.find_element_by_id("id_desc").send_keys("Test modelo")
-        driver.find_element_by_id("id_question_options").click()
-        Select(driver.find_element_by_id("id_question_options")).select_by_visible_text("Preference question")
-        driver.find_element_by_id("id_question_options").click()
-        driver.find_element_by_id("id_options-0-option").click()
-        driver.find_element_by_id("id_options-0-option").clear()
-        driver.find_element_by_id("id_options-0-option").send_keys("intel")
-        driver.find_element_by_id("id_options-0-number").click()
-        driver.find_element_by_id("id_options-0-number").clear()
-        driver.find_element_by_id("id_options-0-number").send_keys("1")
-        driver.find_element_by_id("id_options-1-option").click()
-        driver.find_element_by_id("id_options-1-option").clear()
-        driver.find_element_by_id("id_options-1-option").send_keys("amd")
-        driver.find_element_by_id("id_options-1-number").click()
-        driver.find_element_by_id("id_options-1-number").clear()
-        driver.find_element_by_id("id_options-1-number").send_keys("2")
-        driver.find_element_by_name("_save").click()
-    
-    def is_element_present(self, how, what):
-        try: self.driver.find_element(by=how, value=what)
-        except NoSuchElementException as e: return False
-        return True
-    
-    def is_alert_present(self):
-        try: self.driver.switch_to_alert()
-        except NoAlertPresentException as e: return False
-        return True
-    
-    def close_alert_and_get_its_text(self):
-        try:
-            alert = self.driver.switch_to_alert()
-            alert_text = alert.text
-            if self.accept_next_alert:
-                alert.accept()
-            else:
-                alert.dismiss()
-            return alert_text
-        finally: self.accept_next_alert = True
-    
-    def tearDown(self):
-        self.driver.quit()
-        self.assertEqual([], self.verificationErrors)
-
-# if __name__ == "__main__":
-#     unittest.main()
-
-
-#TEST2 CON SELENIUM MODELO FLOR
-class TestQues(StaticLiveServerTestCase):
-    def setUp(self):
-        self.driver = webdriver.Firefox()
-        self.driver.implicitly_wait(30)
-        self.base_url = "https://www.google.com/"
-        self.verificationErrors = []
-        self.accept_next_alert = True
-    
-    def test_ques(self):
-        driver = self.driver
-        driver.get("http://localhost:8000/admin/login/?next=/admin/")
-        driver.find_element_by_id("id_username").click()
-        driver.find_element_by_id("id_username").clear()
-        driver.find_element_by_id("id_username").send_keys("admin")
-        driver.find_element_by_id("id_password").click()
-        driver.find_element_by_id("id_password").clear()
-        driver.find_element_by_id("id_password").send_keys("flocorlop")
-        driver.find_element_by_xpath("//input[@value='Log in']").click()
-        driver.find_element_by_link_text("Questions").click()
-        driver.find_element_by_class_name("addlink").click()
-        driver.find_element_by_id("id_desc").click()
-        driver.find_element_by_id("id_desc").clear()
-        driver.find_element_by_id("id_desc").send_keys("Test modelo 2")
-        driver.find_element_by_id("id_question_options").click()
-        Select(driver.find_element_by_id("id_question_options")).select_by_visible_text("Preference question 2")
-        driver.find_element_by_id("id_question_options").click()
-        driver.find_element_by_id("id_options-0-option").click()
-        driver.find_element_by_id("id_options-0-option").clear()
-        driver.find_element_by_id("id_options-0-option").send_keys("apple")
-        driver.find_element_by_id("id_options-0-number").click()
-        driver.find_element_by_id("id_options-0-number").clear()
-        driver.find_element_by_id("id_options-0-number").send_keys("1")
-        driver.find_element_by_id("id_options-1-option").click()
-        driver.find_element_by_id("id_options-1-option").clear()
-        driver.find_element_by_id("id_options-1-option").send_keys("android")
-        driver.find_element_by_id("id_options-1-number").click()
-        driver.find_element_by_id("id_options-1-number").clear()
-        driver.find_element_by_id("id_options-1-number").send_keys("2")
-        driver.find_element_by_id("id_options-2-option").click()
-        driver.find_element_by_id("id_options-2-option").clear()
-        driver.find_element_by_id("id_options-2-option").send_keys("microsoft")
-        driver.find_element_by_id("id_options-2-number").click()
-        driver.find_element_by_id("id_options-2-number").clear()
-        driver.find_element_by_id("id_options-2-number").send_keys("3")
-        driver.find_element_by_name("_save").click()
-        #assert de pagina 200 todo
-    
-    def is_element_present(self, how, what):
-        try: self.driver.find_element(by=how, value=what)
-        except NoSuchElementException as e: return False
-        return True
-    
-    def is_alert_present(self):
-        try: self.driver.switch_to_alert()
-        except NoAlertPresentException as e: return False
-        return True
-    
-    def close_alert_and_get_its_text(self):
-        try:
-            alert = self.driver.switch_to_alert()
-            alert_text = alert.text
-            if self.accept_next_alert:
-                alert.accept()
-            else:
-                alert.dismiss()
-            return alert_text
-        finally: self.accept_next_alert = True
-    
-    def tearDown(self):
-        self.driver.quit()
-        self.assertEqual([], self.verificationErrors)
-
-# if __name__ == "__main__":
-#     unittest.main()
